@@ -2,50 +2,53 @@
   "private": true,
   "license": "MIT",
   "scripts": {
-    "prebuild": "npm run clean",
-    "build": "NODE_ENV=production webpack -p",
-    "clean": "rm -rf build",
-    "test": "NODE_ENV=test NODE_PATH=src/client mocha",
+    "build": "run-p build:*",
+    "build:css": "gulp build:css -p",
+    "build:js": "gulp build:js -p",
+    "test": "NODE_PATH=src/client mocha",
     "upgrade": "npm-upgrade",
-    "watch": "webpack-dev-server --hot --inline"
+    "watch": "run-p watch:*",
+    "watch:css": "gulp watch:css",
+    "watch:js": "webpack-dev-server --hot --inline"
   },
   "dependencies": {
     "baobab": "2.3.3",
     "baobab-react": "2.1.1",
-    "bluebird": "3.3.5",
-    "react": "15.0.2",
-    "react-dom": "15.0.2",
-    "react-router": "2.4.0",
-    "superagent": "1.8.3",
+    "bluebird": "3.4.0",
+    "react": "15.1.0",
+    "react-dom": "15.1.0",
+    "react-router": "2.4.1",
+    "superagent": "2.0.0-alpha.3",
     "url-pattern": "1.0.1"
   },
   "devDependencies": {
-    "autoprefixer": "6.3.6",
-    "babel-core": "6.8.0",
     "babel-loader": "6.2.4",
+    "babel-plugin-transform-decorators-legacy": "1.3.4",
     "babel-plugin-transform-react-constant-elements": "6.8.0",
     "babel-plugin-transform-react-inline-elements": "6.8.0",
-    "babel-plugin-transform-runtime": "6.8.0",
-    "babel-preset-es2015": "6.6.0",
+    "babel-plugin-transform-runtime": "6.9.0",
+    "babel-preset-es2015": "6.9.0",
     "babel-preset-react": "6.5.0",
     "babel-preset-react-hmre": "1.1.1",
     "babel-preset-stage-0": "6.5.0",
     "bootstrap": "3.3.6",
-    "css-loader": "0.23.1",
-    "extract-text-webpack-plugin": "1.0.1",
-    "file-loader": "0.8.5",
+    "gulp": "3.9.1",
+    "gulp-autoprefixer": "3.1.0",
+    "gulp-csso": "2.0.0",
+    "gulp-stylus": "2.3.1",
     "html-webpack-plugin": "2.17.0",
     "jsdom": "9.1.0",
     "json-loader": "0.5.4",
+    "minimist": "1.2.0",
     "mocha": "2.4.5",
+    "npm-run-all": "2.1.1",
     "npm-upgrade": "0.6.2",
-    "postcss-loader": "0.9.1",
-    "should": "8.3.1",
+    "null-loader": "0.1.1",
+    "raw-loader": "0.5.1",
+    "should": "8.3.2",
     "style-loader": "0.13.1",
-    "stylus": "0.54.5",
-    "stylus-loader": "2.0.0",
-    "url-loader": "0.5.7",
-    "webpack": "1.13.0",
+    "through": "2.3.8",
+    "webpack": "1.13.1",
     "webpack-dev-server": "1.14.1"
   }
 }
