@@ -1,4 +1,4 @@
-import 'css/index.styl'
+import 'css/index.css'
 import * as actions from 'actions'
 import Baobab from 'baobab'
 import {browserHistory, Router} from 'react-router'
